vm: add helpers to configure and run a VM
This is really just a port of TestRun from the kvm package. It wraps
all the repetitive boilerplate in a call to vm.New(vm.Config{}) and
introduces vm.Loader to configure the VM's memory and registers.

The Run method just runs kvm.Run on the first VCPU in a loop, returning
if the VM halts (KVM_EXIT_HLT). Any other vmexit panics. Eventually
each VCPU will need to run in its own host thread.