hype: build a guest kernel
Eventually hype is going to boot Linux, which means we need to build
a guest kernel. So this adds a lib/linux submodule pointed at linux-
stable and some minimal Makefile scaffolding to build a bzImage.

On amd64, running `make -j $(nproc)` will build
.build/linux/guest/arch/x86/boot/bzImage. The guest kernel config is
currently based on the x86 default config with KVM enabled and all
modules set to "no". To edit it, run `make menuconfig-guest`.