linux: parse a bzImage zeropage (amd64)
When a boot loader loads a bzImage, it parses the first (4K) page of
the file (the zeropage) to figure out what it's dealing with. Then it
uses the information from the zeropage to construct a clean zeropage in
memory.

This adds linux.ParseBzImage, which parses a zeropage into a struct.
(linux.BootParams). This struct is very similar to struct boot_params,
defined in Linux's bootparam.h. But since the C struct is packed, we
can't create a Go struct with exactly the same layout. So instead
BootParams can be BinaryMarshaled and BinaryUnmarshaled to/from the
in-memory layout of struct boot_params.