vm: describe the VM to the Loader
The Loader interface is basically a boot loader. Today there's not much
that it needs to know about the "hardware" it's booting: The only
configurable property (aside from the Loader itself) is the memory size,
which is already known to LoadMemory as len(mem). But soon there will be
multiple VCPUs and virtio devices, and the Loader will need to know
about them too.

Instead of adding multiple params to LoadMemory and LoadVCPU, this
change introduces vm.Info, a struct that describes a configured (but
not yet loaded) VM. An info struct is created a for a new VM after it is
configured, and a copy of the info is passed to the Loader's LoadMemory
and LoadVCPU methods.

There's initially only one field: MemSize.