vm: separate arch-specific setup
The only arch that matters right now is amd64, but there's still tons
of arch-specific stuff that shouldn't be interleaved. So this adds a
vm.Arch interface and a default noop amd64 implementation in the
vm/arch package. For testing, the default implementation can be
overridden by setting Config.Arch.