kvm: expose KVM_EXIT_IO and KVM_EXIT_MMIO data
The kvm_run struct available by mmaping a vcpu fd exposes vmexit data
as a union of anonymous structs. Go hates this, of course. This change
exposes the exit data for the KVM_EXIT_IO and KVM_EXIT_MMIO vmexits as
a Go struct by punning a pointer to the vmexit data field.