etc/linux: enable virtio-mmio in the guest kernel
To boot a guest with virtio-mmio devices we need a guest kernel with
virtio-mmio support. This change enables the virtio-mmio bus in the
guest kernel config, along with support for specifying virtio-mmio
devices on the kernel cmdline like this:

    virtio_mmio.device=0x1000@0xd0000000:5

Now the boot loader can append the VM's virtio-mmio devices to the
guest kernel cmdline. It would be nice to use a devicetree instead.
Can x86 kernels do that?