virtio/virtq: rework the interface, add notify
I originally implemented this in 7d97ab4 based on the spec without using
it. And now that I am, it feels pretty bad, mostly because I have to
deal with notification manually when I release a chain. The rest of the
chain interface doesn't feel great either.

So now when a virtqueue is created it gets an additional notify callback
func, which is called when a chain is released. And a chain has Len(),
Desc(index) and Data(index) methods instead of exposing its internals.