vm: stub virtio-mmio
This is just enough to make a console work. virtio.DeviceHandler is
the device interface, initially only implemented by *virtio.Console.
The VM's mmio.Bus creates a device for each handler, assigning it a
memory area and an IRQ. When a KVM_EXIT_MMIO occurs, the bus handles
it by calling the appropriate device. In the opposite direction,
interrupts are implemented using eventfd/irqfd.

MMIO devices are communicated to the guest by adding entries like
"virtio_mmio.device=%#x@%#x:%d" to the kernel cmdline. It would
be cool to pass a devicetree instead. Not sure if you can do that
with an amd64 kernel.

In the vm_test package, it's now possible to write a test that has
host and guest parts. When TestMain runs on the host, it recompiles
the test binary with -tags guest, which includes vm_guest_test.go,
which is otherwise excluded by a go:build guest directive. Then it
generates an initrd.cpio.gz with the test binary as /init.

The runGuest test helper, which is only available on the host,
boots a Linux guest using the test initrd and passes
-test.run=test.Name() to run the corresponding guest test.

It doesn't have any tests other than a basic end-to-end for the
console. It's undoubtedly very slow. The device handler interface
doesn't support device-specific config registers, which will be
required to implement a block device. The mmio bus is super basic.
There's no provision for devices to be closed when the vm is closed.

But it prints to the terminal, and that's enough for me :)