hype: build a debug initrd
It's not going to be useful to boot Linux unless we can do something
in userland. So this change adds .build/initrd.cpio.gz, based on the
alpine Docker image plus a tiny init script that runs getty. This should
be useful for both automated tests and manual debugging.

The initrd is generated using cmd/docker2cpio, a cursed little command
that reads tar and writes cpio. Run `make .build/initrd.cpio.gz` to
build the archive. Running `make guest` will now build the guest kernel
and the initrd.