virtio: add a block device
This adds a block device with pluggable storage, along with a few basic
storage implementations (memory, file, HTTP). Only reads and optionally
writes are supported: No multiqueue, flush, discard, &c yet.

The handler implementation in this commit is bad and should be replaced.

To support block device config reads, the DeviceHandler interface adds a
ReadConfig method. The mmio bus calls it when the driver reads a device-
specific config register.