vmm: ignore SIGURG
Back in b803cb9 I changed the runloop to ignore EINTR because it was
happening constantly. But we actually need to use EINTR to signal VM
interruption, so this patch stops the errors in a different way. Maybe
worse, because it ignores SIGURG for the entire process.

It turns out KVM_RUN was frequently returning EINTR because of a
pending SIGURG. The Go runtime generates these signals when it is
preempting an uncooperative goroutine. And we're not helping since
we're calling KVM_RUN without dedicating a vcpu thread.